일반 게시판 이미지 추가 및 수정, 삭제 구현

+++ src/main/java/com/kh/pp/plant/model/vo/Plant.java
@@ -9,7 +9,7 @@ import lombok.Value;
 @Builder
 public class Plant {
 	private Long plantNo;
-	private	Integer memberNo;
+	private	Long memberNo;
 	private	String plantName;
 	private	String classification;
 	private	Integer count;
